[BE] Refactor: Migration ApiResponse
- Migration ApiResponse
