+++ BE/src/main/java/com/codesquad/sidedish08/controller/AuthController.java
@@ -2,7 +2,7 @@ package com.codesquad.sidedish08.controller;
 
 import com.codesquad.sidedish08.message.AuthMessages;
 import com.codesquad.sidedish08.message.SuccessMessages;
-import com.codesquad.sidedish08.response.ApiResponse;
+import com.codesquad.sidedish08.model.response.ApiResponse;
 import com.codesquad.sidedish08.service.AuthService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
