+++ BE/src/main/java/com/codesquad/sidedish08/controller/BestDishController.java
@@ -2,13 +2,12 @@ package com.codesquad.sidedish08.controller;
 
 import com.codesquad.sidedish08.message.DishMessages;
 import com.codesquad.sidedish08.message.SuccessMessages;
-import com.codesquad.sidedish08.response.ApiResponse;
+import com.codesquad.sidedish08.model.response.ApiResponse;
 import com.codesquad.sidedish08.service.BasicService;
 import com.codesquad.sidedish08.service.BestDishService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
