+++ BE/src/main/java/com/codesquad/sidedish08/controller/DevController.java
@@ -1,8 +1,7 @@
 package com.codesquad.sidedish08.controller;
 
-import com.codesquad.sidedish08.message.DishMessages;
 import com.codesquad.sidedish08.message.SuccessMessages;
-import com.codesquad.sidedish08.response.ApiResponse;
+import com.codesquad.sidedish08.model.response.ApiResponse;
 import com.codesquad.sidedish08.service.DevService;
 import com.codesquad.sidedish08.util.ResponseUtils;
 import io.swagger.annotations.Api;
