+++ BE/src/main/java/com/codesquad/sidedish08/model/response/ApiResponse.java
@@ -1,4 +1,6 @@
-package com.codesquad.sidedish08.response;
+package com.codesquad.sidedish08.model.response;
+
+import static com.codesquad.sidedish08.util.ResponseUtils.getResultMap;
 
 import java.util.Map;
 
@@ -14,6 +16,10 @@ public class ApiResponse {
 
   public static ApiResponse ok(String message, Map<String, ?> contents) {
     return new ApiResponse(message, contents);
+  }
+
+  public static ApiResponse error(String message, String errorTrace) {
+    return new ApiResponse(message, getResultMap("errorTrace", errorTrace));
   }
 
   public String getMessage() {
